# swap: speed up by preferring a 1MiB blocksize