@justincormack 2019-12-10
Swap files must not have holes, so you must use dd.

On Mon, 9 Dec 2019 at 19:20, Robb Kistler <notifications@github.com> wrote:

> If I add a sync and time it separately, I get the following results. It's
> not so much that the bigger block size is much more efficient, the writes
> are being cached. Is it safe to use fallocate here instead of dd? I vaguely
> remember that fallocate might leave "holes" which swap doesn't like.
>
> time dd if=/dev/zero of=output bs=1048576 count=1024 && time sync
> 1024+0 records in
> 1024+0 records out
> 1073741824 bytes (1.1 GB, 1.0 GiB) copied, 0.571346 s, 1.9 GB/s
>
> real	0m0.572s
> user	0m0.000s
> sys	0m0.572s
>
> real	0m10.289s
> user	0m0.000s
> sys	0m0.000s
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/linuxkit/linuxkit/pull/3441?email_source=notifications&email_token=AADVYPCWL2HOV37C66C5ZXLQX4DGBA5CNFSM4JQVZ5K2YY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOEGMQLPA#issuecomment-563676604>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AADVYPHDXDJUY4X6YS74YP3QX4DGBANCNFSM4JQVZ5KQ>
> .
>